Add files via upload
by means of the equipment available in the laboratory, measurements were made
experimental using an electrical circuit and oscilloscope. The goal is to acquire the data
necessary to identify the system that has as input a VCO, generating two outputs in the form
of sinusoidal signal. This data will include information about the time, the signal of
input and the two output signals obtained.